Refactors entry handlers into separate modules
